refactor(ui): разделён SidebarView на focused sidebar views
SidebarView оставлен root container для SidebarScreenController, AppState и refresh flow
toolbar, source filter menu, content list, rows, refresh overlay и source icon loading вынесены в отдельные файлы
