+++ README.md
@@ -537,7 +537,7 @@
 - [x] `Roadmap TODO Cleanup`: перенести временные inline TODO для будущих OPML import/export flows из runtime UI-кода в roadmap / Post-MVP backlog или заменить их нейтральными disabled actions, чтобы текущий код не хранил незакрытые комментарии по отложенной функциональности;
 - [x] `App Dependencies Test File Physical Split`: после логического split по `@Suite` физически разнести `AppDependenciesTests` на файлы по composition areas и вынести локальные doubles / helpers в focused test support, чтобы targeted test runs не требовали навигации по одному большому файлу;
 - [x] `Reader View Interaction Split`: разрезать `ReaderView` на root container, article chrome/actions, adjacent navigation gestures / overscroll indicators, body block presentation и cached image loading views; оставить root view координатором `ArticleScreenState`, `ArticleScreenActionHandlers` и renderer output;
-- [ ] `Sidebar View Boundary Split`: разделить `SidebarView` на navigation container, toolbar/actions, source filter menu, rows/sections и source icon loading view, чтобы sidebar UI не смешивал routing, custom refresh gesture и favicon cache/network fallback;
+- [x] `Sidebar View Boundary Split`: разделить `SidebarView` на navigation container, toolbar/actions, source filter menu, rows/sections и source icon loading view, чтобы sidebar UI не смешивал routing, custom refresh gesture и favicon cache/network fallback;
 - [ ] `Settings Presentation Boundary Split`: разнести `SettingsScreenPresentationModels` и `SettingsScreenController` на section builders / action handlers по areas: appearance, reading, article list, updates & sync, notifications и storage; root controller должен координировать loading/apply flow, а не хранить всю presentation policy;
 - [ ] `Source Management Service Facade Split`: оставить `DefaultSourceManagementService` публичным facade, но вынести feed preview/discovery, feed command persistence, folder command persistence, move source flow, normalization/uniqueness policy и summary mapping в отдельные focused collaborators;
 - [ ] `App Action Router Area Split`: физически разделить `AppActionRouter` на navigation/Safari actions, source-management actions, refresh actions, cleanup/storage actions и badge actions, сохранив единый app-level action surface поверх `AppState` и services;
